+++ valhalla/odin/maneuver.h
@@ -186,6 +186,13 @@ public:
   bool intersecting_forward_edge() const;
   void set_intersecting_forward_edge(bool intersecting_forward_edge);
 
+  const std::string& verbal_succinct_transition_instruction() const;
+  void set_verbal_succinct_transition_instruction(
+      const std::string& verbal_succinct_transition_instruction);
+  void
+  set_verbal_succinct_transition_instruction(std::string&& verbal_succinct_transition_instruction);
+  bool HasVerbalSuccinctTransitionInstruction() const;
+
   const std::string& verbal_transition_alert_instruction() const;
   void
   set_verbal_transition_alert_instruction(const std::string& verbal_transition_alert_instruction);
@@ -365,6 +372,9 @@ public:
   DirectionsLeg_Maneuver_BssManeuverType bss_maneuver_type() const;
   void set_bss_maneuver_type(DirectionsLeg_Maneuver_BssManeuverType);
 
+  bool has_long_street_name() const;
+  void set_long_street_name(bool has_long_street_name);
+
 #ifdef LOGGING_LEVEL_TRACE
   std::string ToString() const;
 
@@ -404,6 +414,7 @@ protected:
   bool fork_;
   bool begin_intersecting_edge_name_consistency_;
   bool intersecting_forward_edge_;
+  std::string verbal_succinct_transition_instruction_;
   std::string verbal_transition_alert_instruction_;
   std::string verbal_pre_transition_instruction_;
   std::string verbal_post_transition_instruction_;
@@ -434,6 +445,7 @@ protected:
 
   bool has_collapsed_small_end_ramp_fork_;
   bool has_collapsed_merge_maneuver_;
+  bool has_long_street_name_;
 
   ////////////////////////////////////////////////////////////////////////////
   // Transit support